scripts/requirements.txt
openai>=1.47.0
python-slugify>=8.0.4
PyYAML>=6.0.2
requests>=2.31.0
python-frontmatter>=1.1.0
